renamed classes and updated BibliotecaAppTest
