+++ test/com/twu/biblioteca/BibliotecaAppTest.java
@@ -9,56 +9,66 @@ public class BibliotecaAppTest {
     @Test //[#1.2 & #1.3]
     public void displayAvailableBooksCalled() {
         BibliotecaApp bibliotecaApp = new BibliotecaApp();
-        DummyMainMenuInterface dummyMainMenu = new DummyMainMenuInterface();
+        DummyMainMenu dummyMainMenu = new DummyMainMenu();
         String userInput = "1";
 
-        bibliotecaApp.switchOptions(userInput, dummyMainMenu, true);
+        bibliotecaApp.switchOptions(dummyMainMenu, true, userInput);
 
         assertTrue(dummyMainMenu.displayAvailableBooksHasBeenCalled);
     }
 
     @Test //[#1.5]
     public void showInvalidMessageCalled() {
-        DummyMainMenuInterface mainMenuInterface = new DummyMainMenuInterface();
+        BibliotecaApp bibliotecaApp = new BibliotecaApp();
+        DummyMainMenu dummyMainMenu = new DummyMainMenu();
+        String userInput = "a";
 
-        mainMenuInterface.showInvalidMessage();
+        bibliotecaApp.switchOptions(dummyMainMenu, true, userInput);
 
-        assertTrue(mainMenuInterface.showInvalidMessageHasBeenCalled);
+        assertTrue(dummyMainMenu.showInvalidMessageHasBeenCalled);
     }
 
     @Test //[#1.6]
     public void showGoodbyeMessageCalled() {
-        DummyMainMenuInterface mainMenuInterface = new DummyMainMenuInterface();
+        BibliotecaApp bibliotecaApp = new BibliotecaApp();
+        DummyMainMenu dummyMainMenu = new DummyMainMenu();
+        String userInput = "3";
 
-        mainMenuInterface.showGoodbyeMessage();
+        bibliotecaApp.switchOptions(dummyMainMenu, true, userInput);
 
-        assertTrue(mainMenuInterface.showGoodbyeMessageHasBeenCalled);
+        assertTrue(dummyMainMenu.showGoodbyeMessageHasBeenCalled);
     }
 
     @Test //[#1.7]
     public void checkOutBookCalled() {
-        DummyMainMenuInterface mainMenuInterface = new DummyMainMenuInterface();
+        BibliotecaApp bibliotecaApp = new BibliotecaApp();
+        DummyMainMenu dummyMainMenu = new DummyMainMenu();
+        String userInput = ("2" + System.getProperty("line.separator"));
 
-        mainMenuInterface.checkOutBook();
+        bibliotecaApp.switchOptions(dummyMainMenu, true, userInput);
 
-        assertTrue(mainMenuInterface.checkOutBookHasBeenCalled);
+        assertTrue(dummyMainMenu.checkOutBookHasBeenCalled);
     }
 
     @Test //[#1.8]
     public void showSuccessfulCheckOutMessageCalled() {
-        DummyMainMenuInterface mainMenuInterface = new DummyMainMenuInterface();
+        BibliotecaApp bibliotecaApp = new BibliotecaApp();
+        DummyMainMenu dummyMainMenu = new DummyMainMenu ();
+        String userInput = ("2" + System.getProperty("line.separator") + "1" + System.getProperty("line.separator"));
 
-        mainMenuInterface.showSuccessfulCheckOutMessage();
+        bibliotecaApp.switchOptions(dummyMainMenu, true, userInput);
 
-        assertTrue(mainMenuInterface.showSuccessfulCheckOutMessageHasBeenCalled);
+        assertTrue(dummyMainMenu.showSuccessfulCheckOutMessageHasBeenCalled);
     }
 
     @Test //[#1.9]
     public void showUnsuccessfulCheckOutMessageCalled() {
-        DummyMainMenuInterface mainMenuInterface = new DummyMainMenuInterface();
+        BibliotecaApp bibliotecaApp = new BibliotecaApp();
+        DummyMainMenu  dummyMainMenu = new DummyMainMenu ();
+        String userInput = ("2" + System.getProperty("line.separator") + "q" + System.getProperty("line.separator"));
 
-        mainMenuInterface.showUnsuccessfulCheckOutMessage();
+        bibliotecaApp.switchOptions(dummyMainMenu, true, userInput);
 
-        assertTrue(mainMenuInterface.showUnsuccessfulCheckOutMessageHasBeenCalled);
+        assertTrue(dummyMainMenu.showUnsuccessfulCheckOutMessageHasBeenCalled);
     }
 }
